move author name colouring code into its own file
